Move platform tests to their own file and get 100 coverage
