+++ ros_cross_compile/platform.py
@@ -103,10 +103,6 @@ class Platform:
         return self._os_distro
 
     @property
-    def cc_toolchain(self):
-        return self._cc_toolchain
-
-    @property
     def ros_version(self):
         return self._ros_version
 
@@ -123,8 +119,3 @@ class Platform:
     def target_base_image(self) -> str:
         """Name of the base OS Docker image for the target architecture."""
         return self._docker_target_base
-
-    @property
-    def native_base_image(self) -> str:
-        """Name of the base OS Docker image for the host platform architecture."""
-        return self._docker_native_base
