+++ test/test_sysroot_creator.py
@@ -15,7 +15,6 @@
 
 """Unit tests for the `create_cc_sysroot.py` script."""
 
-import getpass
 import os
 from pathlib import Path
 from typing import Tuple
@@ -58,31 +57,6 @@ def setup_mock_sysroot(path: Path) -> Tuple[Path, Path]:
     qemu_binary_mock = qemu_dir / 'qemu'
     qemu_binary_mock.ensure()
     return sysroot_dir, ros_workspace_dir
-
-
-def test_platform_argument_validation():
-    p = Platform('armhf', 'ubuntu', 'dashing')
-    assert p
-
-    with pytest.raises(ValueError):
-        # invalid arch
-        p = Platform('mips', 'ubuntu', 'dashing')
-
-    with pytest.raises(ValueError):
-        # invalid distro
-        p = Platform('armhf', 'ubuntu', 'ardent')
-
-    with pytest.raises(ValueError):
-        # invalid OS
-        p = Platform('armhf', 'rhel', 'dashing')
-
-
-def test_sysroot_image_tag(platform_config):
-    """Make sure the image tag is created correctly."""
-    image_tag = platform_config.sysroot_image_tag
-    test_tag = '{}/{}:latest'.format(getpass.getuser(), str(platform_config))
-    assert isinstance(image_tag, str)
-    assert image_tag == test_tag
 
 
 def test_sysroot_creator_constructor(
@@ -168,35 +142,6 @@ def test_sysroot_creator_tree_additions(platform_config, tmpdir):
     assert (sysroot_dir / 'mixins' / 'index.yaml').exists()
 
 
-def verify_base_docker_images(arch, os, rosdistro, image_name):
-    """Assert correct base image is generated."""
-    platform = Platform(arch, os, rosdistro)
-    assert platform.target_base_image == image_name
-
-
-def test_get_docker_base_image():
-    """Test that the correct base docker image is used for all arguments."""
-    verify_base_docker_images('aarch64', 'ubuntu', 'dashing', 'arm64v8/ubuntu:bionic')
-    verify_base_docker_images('aarch64', 'ubuntu', 'eloquent', 'arm64v8/ubuntu:bionic')
-    verify_base_docker_images('aarch64', 'ubuntu', 'kinetic', 'arm64v8/ubuntu:xenial')
-    verify_base_docker_images('aarch64', 'ubuntu', 'melodic', 'arm64v8/ubuntu:bionic')
-
-    verify_base_docker_images('aarch64', 'debian', 'dashing', 'arm64v8/debian:stretch')
-    verify_base_docker_images('aarch64', 'debian', 'eloquent', 'arm64v8/debian:buster')
-    verify_base_docker_images('aarch64', 'debian', 'kinetic', 'arm64v8/debian:jessie')
-    verify_base_docker_images('aarch64', 'debian', 'melodic', 'arm64v8/debian:stretch')
-
-    verify_base_docker_images('armhf', 'ubuntu', 'dashing', 'arm32v7/ubuntu:bionic')
-    verify_base_docker_images('armhf', 'ubuntu', 'eloquent', 'arm32v7/ubuntu:bionic')
-    verify_base_docker_images('armhf', 'ubuntu', 'kinetic', 'arm32v7/ubuntu:xenial')
-    verify_base_docker_images('armhf', 'ubuntu', 'melodic', 'arm32v7/ubuntu:bionic')
-
-    verify_base_docker_images('armhf', 'debian', 'dashing', 'arm32v7/debian:stretch')
-    verify_base_docker_images('armhf', 'debian', 'eloquent', 'arm32v7/debian:buster')
-    verify_base_docker_images('armhf', 'debian', 'kinetic', 'arm32v7/debian:jessie')
-    verify_base_docker_images('armhf', 'debian', 'melodic', 'arm32v7/debian:stretch')
-
-
 def test_parse_docker_build_output(
         platform_config, tmpdir):
     """Test the SysrootCreator constructor assuming valid path setup."""
@@ -231,14 +176,3 @@ def test_parse_docker_build_output(
     ]
     with pytest.raises(docker.errors.BuildError):
         sysroot_creator._parse_build_output(log_generator_with_errors)
-
-
-def test_docker_py_version():
-    # Explicitly check a known difference between apt and pip versions
-    with pytest.raises(TypeError):
-        # 1.20 (from pip, which we are not using) API has named arguments
-        err = docker.errors.BuildError(reason='problem', build_log='stuff that happened')
-
-    # 1.10 API (from apt which we are using) does not
-    err = docker.errors.BuildError('problem')
-    assert err
